Alert Knap

+++ src/components/pages/user/UserPage.js
@@ -51,56 +51,41 @@ function UserPage() {
       .doc(noteID)
       .delete()
       console.log(noteID)
-}
+    }
 
   return (
-<div className="user-component main-area">
-    
-
-  <div className="grid-x user-information">
-  
-  {accounts.map(account => ( 
-    <div className="cell small-10" key={account.id}>
-    
-      <div className="cell"><h1>{account.name}</h1></div>
-      <div className="cell user-information__text">{account.email}</div>
-      <Link className="cell user-information__text-blue" to={"company/" + account.companyID} >Go til Company page</Link>
-      <div className="cell user-information__text"><BiEdit /></div>
-      {/* <input
-        value={account.fname}
-        onChange={e => {
-          setAccounts(e.target.value);
-        }}
-      /> */}
-      <UserUpdate spell={account} />
-
+  <div className="user-component main-area">
+    <div className="grid-x user-information">
+    {accounts.map(account => ( 
+      <div className="cell small-10" key={account.id}>
+        <div className="cell"><h1>{account.name}</h1></div>
+        <div className="cell user-information__text">{account.email}</div>
+        <Link className="cell user-information__text-blue" to={"company/" + account.companyID} >Go til Company page</Link>
+        <div className="cell user-information__text"><BiEdit /></div>
+        <UserUpdate spell={account} />
+      </div>
+      ))}
+      <div className="cell small-2 user-information__logout"> <button className="user-information__button-text" onClick={() => firebase.auth().signOut()}>Logud <FiLogOut /></button> </div>
     </div>
-    ))}
-    <div className="cell small-2 user-information__logout"> <button className="user-information__button-text" onClick={() => firebase.auth().signOut()}>Logud <FiLogOut /></button> </div>
-    
-  </div>
-  <h1>Bookmarks</h1>
-      {notes.map(note => (
-    <div key={note.id} className="grid-x user-bookmarks">
-    
-    <div className="cell small-12 user-bookmarks__titel">
-    <div className="grid-x">
-      
-      <div className="cell small-10 "><h2>{note.itemTitle}</h2></div>
-      <div className="cell small-2 user-bookmarks__delete">
-        <h2>
-        <RiDeleteBinLine onClick={() => deleteItem(note.id)}/>
-          </h2>
-        </div>
-      </div>  
-    </div>
-    <div className="cell small-12 user-bookmarks__text"><span>Format:</span><br />{note.itemFile}</div>
-    <div className="cell small-12 user-bookmarks__text"><span>Firma skal lige ændres i DBEN :</span><br />{note.itemDesc}</div>
-    <div className="cell small-12 user-bookmarks__text"><span>Description:</span><br />{note.itemDesc}</div>
-    <div className="cell small-12 user-bookmarks__comment "><span>Kommentar :</span><br />{note.userNote}</div>
-    
-    </div>
-    ))}
+    <h1>Bookmarks</h1>
+        {notes.map(note => (
+      <div key={note.id} className="grid-x user-bookmarks">
+      <div className="cell small-12 user-bookmarks__titel">
+      <div className="grid-x">
+        <div className="cell small-10 "><h2>{note.itemTitle}</h2></div>
+        <div className="cell small-2 user-bookmarks__delete">
+          <h2>
+          <RiDeleteBinLine onClick={() => window.confirm(`Are you sure you wish to delete ${note.itemTitle}`) && deleteItem(note.id)}/>
+            </h2>
+          </div>
+        </div>  
+      </div>
+      <div className="cell small-12 user-bookmarks__text"><span>Format:</span><br />{note.itemFile}</div>
+      <div className="cell small-12 user-bookmarks__text"><span>Firma skal lige ændres i DBEN :</span><br />{note.itemDesc}</div>
+      <div className="cell small-12 user-bookmarks__text"><span>Description:</span><br />{note.itemDesc}</div>
+      <div className="cell small-12 user-bookmarks__comment "><span>Kommentar :</span><br />{note.userNote}</div>
+      </div>
+      ))}
   </div>
   );
 }
